+++ visualization/app/codeCharta/state/state.module.ts
@@ -1,5 +1,6 @@
 import angular from "angular"
 // Plop: Append module import here
+import { ColorModeService } from "./store/dynamicSettings/colorMode/colorMode.service"
 import { SecondaryMetricsService } from "./store/appSettings/secondaryMetrics/secondaryMetrics.service"
 import { RecentFilesService } from "./store/dynamicSettings/recentFiles/recentFiles.service"
 import { ColorLabelsService } from "./store/appSettings/colorLabels/colorLabels.service"
@@ -56,10 +57,12 @@ import { ExperimentalFeaturesEnabledService } from "./store/appSettings/enableEx
 import { LayoutAlgorithmService } from "./store/appSettings/layoutAlgorithm/layoutAlgorithm.service"
 import { MaxTreeMapFilesService } from "./store/appSettings/maxTreeMapFiles/maxTreeMapFiles.service"
 import { SharpnessModeService } from "./store/appSettings/sharpnessMode/sharpnessMode.service"
+import { ScreenshotToClipboardEnabledService } from "./store/appSettings/enableClipboard/screenshotToClipboardEnabled.service"
 
 angular
 	.module("app.codeCharta.state", ["app.codeCharta"])
 	// Plop: Append service name here
+	.service(camelCase(ColorModeService.name), ColorModeService)
 	.service(camelCase(SecondaryMetricsService.name), SecondaryMetricsService)
 	.service(camelCase(RecentFilesService.name), RecentFilesService)
 	.service(camelCase(ColorLabelsService.name), ColorLabelsService)
@@ -114,3 +117,4 @@ angular
 	.service(camelCase(MaxTreeMapFilesService.name), MaxTreeMapFilesService)
 	.service(camelCase(SharpnessModeService.name), SharpnessModeService)
 	.service(camelCase(ExperimentalFeaturesEnabledService.name), ExperimentalFeaturesEnabledService)
+	.service(camelCase(ScreenshotToClipboardEnabledService.name), ScreenshotToClipboardEnabledService)