+++ visualization/app/codeCharta/state/store/dynamicSettings/dynamicSettings.splitter.ts
@@ -1,6 +1,7 @@
 import { RecursivePartial, DynamicSettings, CCAction, ColorRange } from "../../../codeCharta.model"
 
 // Plop: Append action splitter import here
+import { splitColorModeAction } from "./colorMode/colorMode.splitter"
 import { splitRecentFilesAction } from "./recentFiles/recentFiles.splitter"
 import { splitSortingOptionAction } from "./sortingOption/sortingOption.splitter"
 import { splitEdgeMetricAction } from "./edgeMetric/edgeMetric.splitter"
@@ -18,6 +19,10 @@ export function splitDynamicSettingsActions(payload: RecursivePartial<DynamicSet
 	const actions: CCAction[] = []
 
 	// Plop: Append action split here
+	if (payload.colorMode !== undefined) {
+		actions.push(splitColorModeAction(payload.colorMode))
+	}
+
 	if (payload.recentFiles !== undefined) {
 		actions.push(splitRecentFilesAction(payload.recentFiles))
 	}