+++ visualization/app/codeCharta/ui/legendPanel/legendPanel.component.spec.ts
@@ -11,6 +11,10 @@ import { FilesService } from "../../state/store/files/files.service"
 import { NodeMetricDataService } from "../../state/store/metricData/nodeMetricData/nodeMetricData.service"
 import { StoreService } from "../../state/store.service"
 import { BlacklistService } from "../../state/store/fileSettings/blacklist/blacklist.service"
+import { AreaMetricService } from "../../state/store/dynamicSettings/areaMetric/areaMetric.service"
+import { HeightMetricService } from "../../state/store/dynamicSettings/heightMetric/heightMetric.service"
+import { EdgeMetricService } from "../../state/store/dynamicSettings/edgeMetric/edgeMetric.service"
+import { setEdgeMetricData } from "../../state/store/metricData/edgeMetricData/edgeMetricData.actions"
 
 describe("LegendPanelController", () => {
 	let legendPanelController: LegendPanelController
@@ -52,6 +56,30 @@ describe("LegendPanelController", () => {
 			expect(ColorRangeService.subscribe).toHaveBeenCalledWith($rootScope, legendPanelController)
 		})
 
+		it("should subscribe to areaMetric", () => {
+			AreaMetricService.subscribe = jest.fn()
+
+			rebuildController()
+
+			expect(AreaMetricService.subscribe).toHaveBeenCalledWith($rootScope, legendPanelController)
+		})
+
+		it("should subscribe to heightMetric", () => {
+			HeightMetricService.subscribe = jest.fn()
+
+			rebuildController()
+
+			expect(HeightMetricService.subscribe).toHaveBeenCalledWith($rootScope, legendPanelController)
+		})
+
+		it("should subscribe to edgeMetric", () => {
+			EdgeMetricService.subscribe = jest.fn()
+
+			rebuildController()
+
+			expect(EdgeMetricService.subscribe).toHaveBeenCalledWith($rootScope, legendPanelController)
+		})
+
 		it("should subscribe to IsAttributeSideBarVisibleService", () => {
 			IsAttributeSideBarVisibleService.subscribe = jest.fn()
 
@@ -86,7 +114,7 @@ describe("LegendPanelController", () => {
 		})
 
 		it("should update _viewModel.maxMetricValue", () => {
-			nodeMetricDataService.getMaxMetricByMetricName = jest.fn(() => 34)
+			nodeMetricDataService.getMaxValueOfMetric = jest.fn(() => 34)
 
 			legendPanelController.onFilesSelectionChanged([])
 
@@ -104,7 +132,7 @@ describe("LegendPanelController", () => {
 		})
 
 		it("should update _viewModel.maxMetricValue", () => {
-			nodeMetricDataService.getMaxMetricByMetricName = jest.fn(() => 34)
+			nodeMetricDataService.getMaxValueOfMetric = jest.fn(() => 34)
 
 			legendPanelController.onFilesSelectionChanged([])
 
@@ -112,9 +140,46 @@ describe("LegendPanelController", () => {
 		})
 	})
 
+	describe("onAreaMetricChanged", () => {
+		it("should update the area metric when it is changed", () => {
+			const newAreaMetric = "new_area_metric"
+
+			legendPanelController.onAreaMetricChanged(newAreaMetric)
+
+			expect(legendPanelController["_viewModel"].areaMetric).toEqual(newAreaMetric)
+		})
+	})
+
+	describe("onHeightMetricChanged", () => {
+		it("should update the height metric when it is changed", () => {
+			const newHeightMetric = "new_height_metric"
+
+			legendPanelController.onHeightMetricChanged(newHeightMetric)
+
+			expect(legendPanelController["_viewModel"].heightMetric).toEqual(newHeightMetric)
+		})
+	})
+
+	describe("onEdgeMetricChanged", () => {
+		it("should update the edge metric and determine if edges exist", () => {
+			const newEdgeMetric = "new_edge_metric"
+
+			storeService.dispatch(setEdgeMetricData([{ name: newEdgeMetric, maxValue: 0, minValue: 0 }]))
+			legendPanelController.onEdgeMetricChanged(newEdgeMetric)
+
+			expect(legendPanelController["_viewModel"].edge).toEqual(newEdgeMetric)
+			expect(legendPanelController["_viewModel"].edgeMetricHasEdge).toBeFalsy()
+
+			storeService.dispatch(setEdgeMetricData([{ name: newEdgeMetric, maxValue: 3, minValue: 1 }]))
+			legendPanelController.onEdgeMetricChanged(newEdgeMetric)
+			expect(legendPanelController["_viewModel"].edge).toEqual(newEdgeMetric)
+			expect(legendPanelController["_viewModel"].edgeMetricHasEdge).toBeTruthy()
+		})
+	})
+
 	describe("onBlackListChanged", () => {
 		it("should update _viewModel.maxMetricValue", () => {
-			nodeMetricDataService.getMaxMetricByMetricName = jest.fn(() => 34)
+			nodeMetricDataService.getMaxValueOfMetric = jest.fn(() => 34)
 
 			legendPanelController.onBlacklistChanged()
 
@@ -124,7 +189,7 @@ describe("LegendPanelController", () => {
 
 	describe("onColorRangeChanged", () => {
 		it("should update the ColorRange when it is changed", () => {
-			const newColorRange: ColorRange = { from: 13, to: 33 }
+			const newColorRange: ColorRange = { from: 13, to: 33, min: 1, max: 10 }
 
 			legendPanelController.onColorRangeChanged(newColorRange)