+++ visualization/app/codeCharta/util/customConfigBuilder.ts
@@ -70,6 +70,7 @@ function initializeAppSettings(target: CustomConfig) {
 		showOnlyBuildingsWithEdges: false,
 		sortingOrderAscending: false,
 		experimentalFeaturesEnabled: false,
+		screenshotToClipboardEnabled: false,
 		layoutAlgorithm: undefined,
 		maxTreeMapFiles: 0,
 		sharpnessMode: undefined,
@@ -107,8 +108,11 @@ function initializeDynamicSettings(target: CustomConfig) {
 		sortingOption: undefined,
 		colorRange: {
 			from: 0,
-			to: 0
+			to: 0,
+			min: 0,
+			max: 0
 		},
+		colorMode: undefined,
 		recentFiles: []
 	}
 }