+++ visualization/setupJest.ts
@@ -4,7 +4,7 @@
 /// <reference path="node_modules/@types/angular-mocks/index.d.ts" />
 
 // e2e tests must not include these - hot fix for issue #2300
-if (process.env.testType !== "e2e") {
+if (process.env.TEST_TYPE !== "e2e") {
 	require("angular")
 	require("angular-mocks")
 	require("jest-preset-angular/setup-jest")